requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Talisman==0.8.1
Flask-Limiter==2.4.0
mysql-connector-python==8.0.26
python-dotenv==0.19.0
pytz==2021.1
Werkzeug==2.0.1
SQLAlchemy==1.4.36
cryptography==3.4.8
requests==2.26.0
reportlab>=3.6.0
Flask-WTF==1.0.0
